added asked questions to profile page
